Fix comment editing.
I wrote this and tested it manually against a single abstract
entry. Everything worked fine. Now that there are more entries,
it didn't work because all the forms had the same dom ids so only
the first entry was getting any action. Added + id to all the ids
being used and things seem to be fine now.