flatten attributes map into columns
I originally thought it would be more flexible to keep attributes in a
map so they're easy to add/remove, but in reality it was a pain in the
butt to have the extra level of indirection in both the Go and
Javascript code. I originally flattened the fields listed in the UI into
columns thinking I'd keep the attributes map around, but really any new
fields should go into the schema so it's gone now too.